move vendor/buford -> buford
