+++ buford/testdata/gopher.png
